+++ cli/app/run.go
@@ -17,17 +17,18 @@ import (
 	"github.com/akamai/cli/pkg/terminal"
 	"github.com/akamai/cli/pkg/tools"
 	"github.com/akamai/cli/pkg/version"
+	"github.com/urfave/cli/v2"
 )
 
-// Run ...
+// Run is the entry point to the CLI
 func Run() int {
 	ctx := context.Background()
 	term := terminal.Color()
+	logger := log.FromContext(ctx)
 
 	var pathErr *os.PathError
 	if err := cleanupUpgrade(); err != nil && errors.As(err, &pathErr) && pathErr.Err != syscall.ENOENT {
-		term.WriteErrorf("Unable to remove old executable: %s", err.Error())
-		return 7
+		logger.Debugf("Unable to remove old executable: %s", err.Error())
 	}
 
 	if err := os.Setenv("AKAMAI_CLI", "1"); err != nil {
@@ -66,11 +67,11 @@ func Run() int {
 		term.WriteErrorf("Unable to export required envs: %s", err.Error())
 	}
 
-	cli := app.CreateApp(ctx)
-	ctx = log.SetupContext(ctx, cli.Writer)
+	cliApp := app.CreateApp(ctx)
+	ctx = log.SetupContext(ctx, cliApp.Writer)
 
 	cmds := commands.CommandLocator(ctx)
-	cli.Commands = cmds
+	cliApp.Commands = cmds
 
 	if err := firstRun(ctx); err != nil {
 		return 5
@@ -80,21 +81,78 @@ func Run() int {
 		term.WriteError(err.Error())
 	}
 
-	if err := cli.RunContext(ctx, os.Args); err != nil {
+	// check command collision
+	if err := findCollisions(cliApp.Commands, os.Args); err != nil {
+		term.WriteError(err)
+		return 7
+	}
+
+	if err := cliApp.RunContext(ctx, os.Args); err != nil {
 		return 6
 	}
 
 	return 0
 }
 
-func cleanupUpgrade() error {
-	oldFilename := os.Args[0]
-	if strings.HasSuffix(strings.ToLower(oldFilename), ".exe") {
-		oldFilename = fmt.Sprintf(".%s.old", oldFilename)
-	} else {
-		oldFilename = fmt.Sprintf(".%s.exe.old", oldFilename)
+func findCollisions(availableCmds []*cli.Command, args []string) error {
+	if len(args) > 1 {
+		// check names and aliases
+
+		// for some built in commands, we need to check their first parameter (args[2])
+		metaCmds := []string{"help", "uninstall", "update"}
+		for _, c := range metaCmds {
+			if c == args[1] && len(args) > 2 {
+				if err := findDuplicate(availableCmds, args[2]); err != nil {
+					return err
+				}
+
+				return nil
+			}
+		}
+
+		// rest of commands: we need to check the first parameter (args[1])
+		if err := findDuplicate(availableCmds, args[1]); err != nil {
+			return err
+		}
 	}
-	return os.Remove(oldFilename)
+
+	return nil
+}
+
+func findDuplicate(availableCmds []*cli.Command, cmdName string) error {
+	matching := make([]string, 0, 2)
+	for _, cmd := range availableCmds {
+		// match with command name
+		if cmd.Name == cmdName {
+			matching = append(matching, cmd.Name)
+			continue
+		}
+		// match with command aliases
+		for _, alias := range cmd.Aliases {
+			if alias == cmdName {
+				matching = append(matching, cmdName)
+				break
+			}
+		}
+
+	}
+
+	if len(matching) > 1 {
+		return fmt.Errorf("this command is ambiguous, please use prefix of the package which should be used (i.e. custom/command): %s", cmdName)
+	}
+
+	return nil
+}
+
+func cleanupUpgrade() error {
+	filename := filepath.Base(os.Args[0])
+	var oldExe string
+	if strings.HasSuffix(strings.ToLower(filename), ".exe") {
+		oldExe = fmt.Sprintf(".%s.old", filename)
+	} else {
+		oldExe = fmt.Sprintf(".%s.exe.old", filename)
+	}
+	return os.Remove(filepath.Join(filepath.Dir(os.Args[0]), oldExe))
 }
 
 func checkUpgrade(ctx context.Context) {